builds

+++ all_seaing_perception/include/all_seaing_perception/factor_graph_slam.hpp
@@ -165,8 +165,8 @@ private:
     float m_imu_xy_noise, m_imu_theta_noise;
     float m_update_gps_xy_uncertainty, m_update_odom_theta_uncertainty;
     int m_num_obj, m_num_banners, m_num_poses;
-    Eigen::Vector3f m_robot_pos_mean;//obstacle map
-    Eigen::Matrix3f m_robot_pos_cov;//covariance matrix
+    Eigen::Vector3d m_robot_pos_mean;//obstacle map
+    Eigen::Matrix3d m_robot_pos_cov;//covariance matrix
     bool m_first_state, m_got_local_frame, m_got_nav, m_got_odom, m_rotate_odom;
     bool m_shouldnt_gps_pred;
     bool m_match_numbers_indicators;
